+++ database/migrations/2014_10_12_100000_create_password_resets_table.php
@@ -15,7 +15,7 @@ class CreatePasswordResetsTable extends Migration
     {
         Schema::create('password_resets', function (Blueprint $table) {
             $table->string('email')->index();
-            $table->string('token');
+            $table->string('token')->index();
             $table->timestamp('created_at')->nullable();
         });
     }
@@ -27,6 +27,6 @@ class CreatePasswordResetsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('password_resets');
+        if(env('APP_ENV') === 'local' || env('APP_ENV') === 'testing') Schema::dropIfExists('password_resets');
     }
 }
